+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## Version 0.1.0 (2023/08/17)
+## Version 0.1.0 (2023/08/21)
 
 - Init
+- Added server prompt sending
+- Added userscript prompting
+- Added WS again to communicate